cleanup: refactor MonsterContainer.tsx
